Initial contribution of the clangd contexts API (#1)
Adds the '@eclipse-cdt-cloud/clangd-contexts' API for managing clangd
contexts in C/C++ projects. Includes an example CLI tool and VS Code
extension showcasing the concepts as well as an example workspace.

- Javascript library for manipulation of .clangd configuration files
  to support multiple contexts within a C/C++ project
- API usage example in the form of a CLI tool
- API usage example in the form of a VS Code extension
- example CMake project showcasing multiple clangd contexts
- also refactors Mocha configuration for deprecation warnings

This contribution has been originally proposed as a contribution to Eclipse
Theia; see https://github.com/eclipse-theia/theia-cpp-extensions/pull/143.
However, after discussing with the authors (Christian Damus and Tobias
Ortmayr, who developed this component on behalf of EclipseSource and
STMicroelectronics) and the project leads of Eclipse Theia and CDT.cloud, we
decided that CDT.cloud is a better fit for this contribution.

Contributed on behalf of STMicroelectronics.

Co-authored-by: Tobias Ortmayr <tortmayr@eclipsesource.com>
Co-authored-by: Philip Langer <planger@eclipsesource.com>
Signed-off-by: Christian W. Damus <cdamus.ext@eclipsesource.com>
Signed-off-by: Philip Langer <planger@eclipsesource.com>